# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/equipment/{id}": {
            "delete": {
                "description": "get equipment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a equipment entity by ID",
                "operationId": "delete-equipment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Equipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/equipments": {
            "get": {
                "description": "list equipment entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List equipment entities",
                "operationId": "list-equipment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Equipment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create equipment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create equipment",
                "operationId": "create-equipment",
                "parameters": [
                    {
                        "description": "Equipment entity",
                        "name": "equipment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Equipment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Equipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/equipments/{id}": {
            "get": {
                "description": "get equipment by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a equipment entity by ID",
                "operationId": "get-equipment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Equipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Equipment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/nearbyplace/{id}": {
            "delete": {
                "description": "get nearbyplace by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a nearbyplace entity by ID",
                "operationId": "delete-nearbyplace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NearbyPlace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/nearbyplaces": {
            "get": {
                "description": "list nearbyplace entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List nearbyplace entities",
                "operationId": "list-nearbyplace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.NearbyPlace"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create nearbyplace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create nearbyplace",
                "operationId": "create-nearbyplace",
                "parameters": [
                    {
                        "description": "NearbyPlace entity",
                        "name": "nearbyplace",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.NearbyPlace"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.NearbyPlace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/nearbyplaces/{id}": {
            "get": {
                "description": "get nearbyplace by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a nearbyplace entity by ID",
                "operationId": "get-nearbyplace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NearbyPlace ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.NearbyPlace"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Equipment": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EquipmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EquipmentEdges"
                },
                "equipment": {
                    "description": "Equipment holds the value of the \"equipment\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.EquipmentEdges": {
            "type": "object",
            "properties": {
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.Facility": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FacilityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FacilityEdges"
                },
                "facility": {
                    "description": "Facility holds the value of the \"facility\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.FacilityEdges": {
            "type": "object",
            "properties": {
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.NearbyPlace": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NearbyPlaceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NearbyPlaceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "placename": {
                    "description": "Placename holds the value of the \"placename\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NearbyPlaceEdges": {
            "type": "object",
            "properties": {
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.Quantity": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the QuantityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.QuantityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "quantity": {
                    "description": "Quantity holds the value of the \"quantity\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.QuantityEdges": {
            "type": "object",
            "properties": {
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "ent.Room": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoomEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "roomprice": {
                    "description": "Roomprice holds the value of the \"roomprice\" field.",
                    "type": "integer"
                },
                "roomtypename": {
                    "description": "Roomtypename holds the value of the \"roomtypename\" field.",
                    "type": "string"
                }
            }
        },
        "ent.RoomEdges": {
            "type": "object",
            "properties": {
                "equipments": {
                    "description": "Equipments holds the value of the equipments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Equipment"
                    }
                },
                "facilities": {
                    "description": "Facilities holds the value of the facilities edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Facility"
                    }
                },
                "nearbyplace": {
                    "description": "Nearbyplace holds the value of the nearbyplace edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.NearbyPlace"
                    }
                },
                "quantity": {
                    "description": "Quantity holds the value of the quantity edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Quantity"
                },
                "staytype": {
                    "description": "Staytype holds the value of the staytype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StayType"
                }
            }
        },
        "ent.StayType": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StayTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StayTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "staytype": {
                    "description": "Staytype holds the value of the \"staytype\" field.",
                    "type": "string"
                }
            }
        },
        "ent.StayTypeEdges": {
            "type": "object",
            "properties": {
                "room": {
                    "description": "Room holds the value of the room edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Room"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}